[ADJUST] Design and Flow
